requirements.txt
pytest==8.3.2
pytest-html==4.1.1
pytest-rerunfailures==9.1.1
pytest-base-url==1.4.2
click==7.1.2
poium==1.5.2
allure-pytest=2.14.2